+++ src/app/cart-list/index.ts
@@ -1,3 +1,4 @@
 export * from './components';
 export * from './services/cart.service';
 export * from './models/cart-item.model';
+export * from './services/cart-observable.service';