+++ src/app/cart-list/models/cart-item.model.ts
@@ -2,6 +2,7 @@ import { ProductModel } from "src/app/shared";
 
 export class CartItemModel {
     constructor(
+        public id: number,
         public product: ProductModel,
         public quantity: number
     ) { }