+++ src/app/core/guards/is-empty-cart.guard.ts
@@ -1,9 +1,12 @@
 import { inject } from '@angular/core';
 import { CanActivateFn } from '@angular/router';
+import { map } from 'rxjs';
 
 import { CartService } from 'src/app/cart-list';
 
 export const isEmptyCartGuard: CanActivateFn = (route, state) => {
   const cartService = inject(CartService);
-  return !cartService.isEmptyCart;
+  return cartService.getProducts().pipe(
+    map(items => !cartService.isEmptyCart(items))
+  );
 };