+++ src/app/core/index.ts
@@ -9,3 +9,4 @@ export * from './models/config.model';
 export * from './guards/can-activate-auth.guard';
 export * from './guards/can-activate-child-auth.guard';
 export * from './services/auth.service';
+export * from './services/app-settings.service';