+++ src/app/product-list/components/product-view/product-view.component.ts
@@ -1,23 +1,24 @@
-import { Component, Input, OnInit } from '@angular/core';
-import { take, tap } from 'rxjs';
+import { Component, Input, OnDestroy, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 
 import { ProductModel } from 'src/app/shared';
-import { ProductService } from '../../services/product.service';
 import { CartService } from 'src/app/cart-list';
+import { ProductsPromiseService } from '../..';
+import { Subscription, tap } from 'rxjs';
 
 @Component({
   selector: 'app-product-view',
   templateUrl: './product-view.component.html',
   styleUrls: ['./product-view.component.css']
 })
-export class ProductViewComponent implements OnInit {
+export class ProductViewComponent implements OnInit, OnDestroy {
 
   @Input() productID!: string;
   product!: ProductModel;
+  private sub!: Subscription;
 
   constructor(
-    private productsService: ProductService,
+    private productService: ProductsPromiseService,
     private router: Router,
     private cartService: CartService
   ) { }
@@ -25,17 +26,21 @@ export class ProductViewComponent implements OnInit {
   ngOnInit(): void {
     this.product = new ProductModel();
 
-    this.productsService.getProductById(+this.productID).pipe(
-      tap((productModel: ProductModel | undefined) => {
-        this.product = productModel ?? {};
-      }),
-      take(1)
-    ).subscribe();
+    this.productService.getProduct(+this.productID)
+      .then((productModel: ProductModel) => {
+        this.product = productModel;
+      })
+      .catch(err => console.log(err));
+  }
+
+  ngOnDestroy(): void {
+    this.sub?.unsubscribe();
   }
 
   onAddToCart() {
-    this.cartService.addProduct(this.product);
-    this.onGoBack();
+    this.sub = this.cartService.addProduct(this.product).pipe(
+      tap(() => this.onGoBack())
+    ).subscribe()
   }
 
   onGoBack(): void {