+++ src/app/product-list/index.ts
@@ -1,2 +1,3 @@
 export * from './components';
 export * from './services/product.service';
+export * from './services/products-promise.service';