+++ src/app/shared/shared.module.ts
@@ -5,6 +5,7 @@ import { HighlightDirective } from './directives/highlight.directive';
 import { BorderDirective } from './directives/border.directive';
 import { OrderByPipe } from './pipes/order-by.pipe';
 import { FormsModule } from '@angular/forms';
+import { HttpClientModule } from '@angular/common/http';
 
 @NgModule({
   declarations: [
@@ -14,14 +15,16 @@ import { FormsModule } from '@angular/forms';
   ],
   imports: [
     CommonModule,
-    FormsModule
+    FormsModule,
+    HttpClientModule
   ],
   exports: [
     HighlightDirective,
     BorderDirective,
     OrderByPipe,
     CommonModule,
-    FormsModule
+    FormsModule,
+    HttpClientModule
   ]
 })
 export class SharedModule { }